pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI application with comprehensive testing"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app" },
    { include = "tests" },
]

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.108.0"
sqlalchemy = "^1.4.23"
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
email-validator = "^2.1.0"
aioredis = "^2.0.0"
httpx = "^0.19.0"
pytest-cov = "^2.12.1"
aiosqlite = "^0.17.0"
pydantic-settings = "^2.0.3"
asyncpg = "^0.29.0"
redis = "^5.0.1"
prometheus-client = "^0.19.0"
psutil = "^5.9.7"
fastapi-mail = "^1.4.1"
slowapi = "^0.1.8"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
python_paths = ["."]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.pylance]
reportMissingImports = "warning"
pythonPath = "."
extraPaths = [".", "tests"] 